Handle null values

+++ CarStack/Repositories/CarRepository.cs
@@ -29,6 +29,9 @@ public sealed class CarRepository(CarDBContext context) : ICarRepository
     public async Task Update(Car car)
     {
         var existingCar = await _context.Cars.FindAsync(car.Id);
+        if (existingCar is null)
+            throw new InvalidOperationException($"Samochód o ID {car.Id} nie został znaleziony.");
+        
         _context.Entry(existingCar).CurrentValues.SetValues(car);
         await _context.SaveChangesAsync();
     }
@@ -36,6 +39,9 @@ public sealed class CarRepository(CarDBContext context) : ICarRepository
     public async Task Delete(int carId)
     {
         var carToDelete = await _context.Cars.FindAsync(carId);
+        if (carToDelete is null)
+            throw new InvalidOperationException($"Samochód o ID {carId} nie został znaleziony.");
+
         _context.Cars.Remove(carToDelete);
         await _context.SaveChangesAsync();
     }
