+++ CarStack/Repositories/UserRepository.cs
@@ -26,6 +26,9 @@ public sealed class UserRepository(CarDBContext context) : IUserRepository
     public async Task Update(CarStackUser user)
     {
         var existingUser = await _context.Users.FindAsync(user.Id);
+        if (existingUser is null)
+            throw new InvalidOperationException($"Użytkownik o ID {user.Id} nie został znaleziony.");
+        
         _context.Entry(existingUser).CurrentValues.SetValues(user);
         await _context.SaveChangesAsync();
     }
@@ -33,6 +36,9 @@ public sealed class UserRepository(CarDBContext context) : IUserRepository
     public async Task Delete(int userId)
     {
         var userToDelete = await _context.Users.FindAsync(userId);
+        if (userToDelete is null)
+            throw new InvalidOperationException($"Użytkownik o ID {userId} nie został znaleziony.");
+
         _context.Users.Remove(userToDelete);
         await _context.SaveChangesAsync();
     }
